+++ iOS-Naduan/Source/Presentation/AuthScene/SignUpView/Controllers/AgreeTermViewController.swift
@@ -36,21 +36,7 @@ class AgreeTermViewController: UIViewController {
   // MARK: - View Properties
   private let titleLabel = SignUpTitleLabel()
   
-  private let nextFlowButton: UIButton = {
-    var container = AttributeContainer()
-    container.font = UIFont.pretendardFont(weight: .bold, size: 18)
-    
-    let inset = NSDirectionalEdgeInsets(top: 10, leading: .zero, bottom: 40, trailing: .zero)
-    var configuration = UIButton.Configuration.daunStyle(with: .basic)
-    configuration.background.cornerRadius = .zero
-    configuration.contentInsets = inset
-    configuration.setTitle(to: TextConstants.nextFlow, with: container)
-    
-    let button = UIButton(configuration: configuration)
-    button.automaticallyUpdatesConfiguration = false
-    button.isEnabled = false
-    return button
-  }()
+  private let nextFlowButton = SignUpNextButton(title: TextConstants.nextFlow)
   
   private let tableView: UITableView = {
     let tableView = UITableView(frame: .zero, style: .plain)
