+++ iOS-Naduan/Source/Presentation/AuthScene/SignUpView/Views/SignUpNextButton.swift
@@ -1,0 +1,23 @@
+//
+//  SignUpNextButton.swift
+//  iOS-Naduan
+//
+//  Copyright (c) 2024 Minii All rights reserved.
+
+import UIKit
+
+class SignUpNextButton: UIButton {
+  convenience init(title: String) {
+    var container = AttributeContainer()
+    container.font = UIFont.pretendardFont(weight: .bold, size: 18)
+    
+    let inset = NSDirectionalEdgeInsets(top: 10, leading: .zero, bottom: 40, trailing: .zero)
+    var configuration = UIButton.Configuration.daunStyle(with: .basic)
+    configuration.background.cornerRadius = .zero
+    configuration.contentInsets = inset
+    configuration.setTitle(to: title, with: container)
+    self.init(configuration: configuration)
+    
+    automaticallyUpdatesConfiguration = false
+  }
+}
